+++ lib/Notification/Notifier.php
@@ -8,6 +8,7 @@ use OCA\NcRoomba\AppInfo\Application;
 use OCP\IURLGenerator;
 use OCP\L10N\IFactory;
 use OCP\Notification\INotification;
+use OCP\Notification\UnknownNotificationException;
 use OCP\Notification\INotifier;
 
 class Notifier implements INotifier
@@ -31,7 +32,7 @@ class Notifier implements INotifier
 	public function prepare(INotification $notification, string $languageCode): INotification
 	{
 		if ($notification->getApp() !== Application::APP_ID) {
-			throw new \InvalidArgumentException();
+			throw new UnknownNotificationException();
 		}
 
 		$l = $this->l10nFactory->get(Application::APP_ID, $languageCode);
@@ -59,7 +60,7 @@ class Notifier implements INotifier
 				$notification->setParsedSubject($l->t('%1$s battery low (%2$s%%)', [$robot, $pct]));
 				break;
 			default:
-				throw new \InvalidArgumentException();
+				throw new UnknownNotificationException();
 		}
 
 		$notification->setIcon(